Fu 37 fenix administrators management q&a

+++ furms-web-ui/src/main/java/io/imunity/furms/ui/views/fenix/administrators/FenixAdministratorsView.java
@@ -5,20 +5,6 @@
 
 package io.imunity.furms.ui.views.fenix.administrators;
 
-import static com.vaadin.flow.component.button.ButtonVariant.LUMO_TERTIARY;
-import static com.vaadin.flow.component.icon.VaadinIcon.ANGLE_DOWN;
-import static com.vaadin.flow.component.icon.VaadinIcon.ANGLE_RIGHT;
-import static com.vaadin.flow.component.icon.VaadinIcon.PAPERPLANE;
-import static com.vaadin.flow.component.icon.VaadinIcon.SEARCH;
-import static com.vaadin.flow.component.icon.VaadinIcon.TRASH;
-import static io.imunity.furms.ui.utils.VaadinExceptionHandler.handleExceptions;
-import static java.util.stream.Collectors.toList;
-
-import java.util.Collections;
-import java.util.List;
-import java.util.Optional;
-import java.util.function.Supplier;
-
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.button.Button;
 import com.vaadin.flow.component.grid.ColumnTextAlign;
@@ -27,23 +13,26 @@ import com.vaadin.flow.component.grid.GridVariant;
 import com.vaadin.flow.component.html.Div;
 import com.vaadin.flow.component.html.Span;
 import com.vaadin.flow.component.icon.VaadinIcon;
-import com.vaadin.flow.component.notification.Notification;
 import com.vaadin.flow.component.orderedlayout.FlexComponent;
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
-import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import com.vaadin.flow.component.textfield.TextField;
-import com.vaadin.flow.data.renderer.ComponentRenderer;
+import com.vaadin.flow.data.provider.ListDataProvider;
 import com.vaadin.flow.data.value.ValueChangeMode;
 import com.vaadin.flow.router.Route;
-
 import io.imunity.furms.domain.users.User;
 import io.imunity.furms.spi.users.UsersDAO;
-import io.imunity.furms.ui.components.FurmsViewComponent;
-import io.imunity.furms.ui.components.GridActionMenu;
-import io.imunity.furms.ui.components.PageTitle;
-import io.imunity.furms.ui.components.SparseGrid;
-import io.imunity.furms.ui.components.ViewHeaderLayout;
+import io.imunity.furms.ui.components.*;
 import io.imunity.furms.ui.views.fenix.menu.FenixAdminMenu;
+
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
+import java.util.function.Supplier;
+
+import static com.vaadin.flow.component.button.ButtonVariant.LUMO_TERTIARY;
+import static com.vaadin.flow.component.icon.VaadinIcon.*;
+import static io.imunity.furms.ui.utils.VaadinExceptionHandler.handleExceptions;
+import static java.util.stream.Collectors.toList;
 
 @Route(value = "fenix/admin/administrators", layout = FenixAdminMenu.class)
 @PageTitle(key = "view.fenix-admin.administrators.page.title")
@@ -54,8 +43,10 @@ public class FenixAdministratorsView extends FurmsViewComponent {
 		this.usersDAO = usersDAO;
 
 		Grid<UserViewModel> grid = createGrid(loadUsers(usersDAO::getAdminUsers));
-		HorizontalLayout inviteUserLayout = createInviteUserLayout(grid);
-		HorizontalLayout searchLayout = createSearchFilterLayout(grid);
+		TextField emailTextField = createEmailField();
+		Button inviteButton = createInviteButton(grid, emailTextField);
+		HorizontalLayout inviteUserLayout = createInviteUserLayout(emailTextField, inviteButton);
+		HorizontalLayout searchLayout = createSearchFilterLayout(grid, inviteButton);
 
 		ViewHeaderLayout headerLayout = new ViewHeaderLayout(getTranslation("view.fenix-admin.administrators.header"), 
 				inviteUserLayout);
@@ -70,9 +61,20 @@ public class FenixAdministratorsView extends FurmsViewComponent {
 			.collect(toList());
 	}
 
-	private HorizontalLayout createInviteUserLayout(Grid<UserViewModel> grid) {
+	private HorizontalLayout createInviteUserLayout(TextField emailTextField, Button inviteButton) {
+		HorizontalLayout horizontalLayout = new HorizontalLayout(emailTextField, inviteButton);
+		horizontalLayout.setJustifyContentMode(FlexComponent.JustifyContentMode.END);
+		horizontalLayout.setSpacing(true);
+		return horizontalLayout;
+	}
+
+	private TextField createEmailField() {
 		TextField emailTextField = new TextField();
 		emailTextField.setPlaceholder(getTranslation("view.fenix-admin.administrators.field.invite"));
+		return emailTextField;
+	}
+
+	private Button createInviteButton(Grid<UserViewModel> grid, TextField emailTextField) {
 		String inviteLabel = getTranslation("view.fenix-admin.administrators.button.invite");
 		Button inviteButton = new Button(inviteLabel, PAPERPLANE.create());
 		inviteButton.addClickListener(event -> {
@@ -84,17 +86,15 @@ public class FenixAdministratorsView extends FurmsViewComponent {
 					user -> {
 						handleExceptions(() -> usersDAO.addFenixAdminRole(user.id));
 						grid.setItems(loadUsers(usersDAO::getAdminUsers));
+						emailTextField.clear();
 						},
-					() -> Notification.show(getTranslation("view.fenix-admin.administrators.error.validation.field.invite"))
+					() -> showErrorNotification(getTranslation("view.fenix-admin.administrators.error.validation.field.invite"))
 				);
 		});
-		HorizontalLayout horizontalLayout = new HorizontalLayout(emailTextField, inviteButton);
-		horizontalLayout.setJustifyContentMode(FlexComponent.JustifyContentMode.END);
-		horizontalLayout.setSpacing(true);
-		return horizontalLayout;
+		return inviteButton;
 	}
 
-	private HorizontalLayout createSearchFilterLayout(Grid<UserViewModel> grid) {
+	private HorizontalLayout createSearchFilterLayout(Grid<UserViewModel> grid, Button inviteButton) {
 		TextField textField = new TextField();
 		textField.setValueChangeMode(ValueChangeMode.EAGER);
 		textField.setPlaceholder(getTranslation("view.fenix-admin.administrators.field.search"));
@@ -109,6 +109,9 @@ public class FenixAdministratorsView extends FurmsViewComponent {
 				)
 				.collect(toList());
 			grid.setItems(filteredUsers);
+			//TODO This is a work around to fix disappearing text cursor
+			inviteButton.focus();
+			textField.focus();
 		});
 
 		HorizontalLayout search = new HorizontalLayout(textField);
@@ -127,25 +130,25 @@ public class FenixAdministratorsView extends FurmsViewComponent {
 		Grid<UserViewModel> grid = new SparseGrid<>(UserViewModel.class);
 		grid.getStyle().set("word-wrap", "break-word");
 		grid.addThemeVariants(GridVariant.LUMO_WRAP_CELL_CONTENT);
-		grid.addComponentColumn(c -> new Div(c.icon, new Span(c.firstName)))
+		grid.addComponentColumn(c -> new Div(c.icon, new Span(c.firstName + " " + c.lastName)))
 			.setHeader(getTranslation("view.fenix-admin.administrators.grid.column.1"));
-		grid.addColumn(c -> c.lastName).setHeader(getTranslation("view.fenix-admin.administrators.grid.column.2"));
-		grid.addColumn(c -> c.email).setHeader(getTranslation("view.fenix-admin.administrators.grid.column.3"));
-		grid.addColumn(c -> "Active").setHeader(getTranslation("view.fenix-admin.administrators.grid.column.4"));
+		grid.addColumn(c -> c.email).setHeader(getTranslation("view.fenix-admin.administrators.grid.column.2"));
+		grid.addColumn(c -> "Active").setHeader(getTranslation("view.fenix-admin.administrators.grid.column.3"));
 		grid.addComponentColumn(c -> {
 			HorizontalLayout horizontalLayout = new HorizontalLayout(addMenu(grid, c.id));
 			horizontalLayout.setJustifyContentMode(FlexComponent.JustifyContentMode.END);
 			return horizontalLayout;
 		})
-			.setHeader(getTranslation("view.fenix-admin.administrators.grid.column.3"))
+			.setHeader(getTranslation("view.fenix-admin.administrators.grid.column.4"))
 			.setTextAlign(ColumnTextAlign.END);
-		grid.setItemDetailsRenderer(new ComponentRenderer<>(
-			data -> {
-				VerticalLayout layout = new VerticalLayout();
-				layout.add(new Span("Example Data"), new Span("Additional Data"));
-				return layout;
-			})
-		);
+		//TODO for now we do not have user data since to show
+//		grid.setItemDetailsRenderer(new ComponentRenderer<>(
+//			data -> {
+//				VerticalLayout layout = new VerticalLayout();
+//				layout.add(new Span("Example Data"), new Span("Additional Data"));
+//				return layout;
+//			})
+//		);
 		grid.addItemClickListener(event -> {
 			event.getItem().icon = grid.isDetailsVisible(event.getItem()) ? ANGLE_DOWN.create() : ANGLE_RIGHT.create();
 			grid.getDataProvider().refreshItem(event.getItem());
@@ -157,10 +160,14 @@ public class FenixAdministratorsView extends FurmsViewComponent {
 	private Component addMenu(Grid<UserViewModel> grid, String id) {
 		GridActionMenu contextMenu = new GridActionMenu();
 		
-		String deleteLabel = getTranslation("view.fenix-admin.administrators.context.menu.delete");
-		contextMenu.addItem(addMenuButton(deleteLabel, TRASH), event -> {
-			handleExceptions(() -> usersDAO.removeFenixAdminRole(id));
-			grid.setItems(loadUsers(usersDAO::getAdminUsers));
+		String deleteLabel = getTranslation("view.fenix-admin.administrators.context.menu.remove");
+		contextMenu.addItem(addMenuButton(deleteLabel, MINUS_CIRCLE), event -> {
+			if(((ListDataProvider<UserViewModel>) grid.getDataProvider()).getItems().size() > 1) {
+				handleExceptions(() -> usersDAO.removeFenixAdminRole(id));
+				grid.setItems(loadUsers(usersDAO::getAdminUsers));
+			}
+			else
+				showErrorNotification(getTranslation("view.fenix-admin.administrators.error.validation.remove"));
 		});
 		getContent().add(contextMenu);
 		return contextMenu.getTarget();
